refactor(mcp): satisfy the module-size ratchet — extract lateral + session-seed surfaces
The P1/P4 features grew three grandfathered modules past their #1797
budgets. Per the gate's own instruction, the new code moves to new
modules — and the extraction pays the ratchet forward:

- LateralThinkHandler moves wholesale to mcp/tools/lateral_think_handler.py
  (re-exported from evaluation_handlers so every import path keeps
  working); evaluation_handlers drops to 1728 lines, under the universal
  2000 cap, so its GRANDFATHERED entry is deleted (policy only tightens)
- decision-mode contracts, task/evidence prompt blocks, and the
  mode/research schema fragments consolidate in mcp/tools/lateral_decision.py
- the interview-less seed MCP surface moves to
  mcp/tools/session_seed_handler.py; authoring_handlers keeps a
  three-line delegation
- remaining single-digit overages closed by tightening the descriptions
  the features added, not by touching pre-existing prose

module-size: OK (619 modules, cap 2000, 25 grandfathered — was 26).

Refs #2278
