+++ src/ouroboros/mcp/tools/subagent.py
@@ -1865,15 +1865,11 @@ def build_lateral_multi_subagent(
         problem_context: Description of the stuck situation.
         current_approach: What has been tried and isn't working.
         failed_attempts: Previous failed approaches shared across all panes.
-        mode: ``"unstuck"`` (default) keeps the historical break-the-stagnation
-              task block. ``"decision"`` reframes each persona as an advocate in
-              a decision advisory: position, strongest argument, strongest
-              counter-argument to its own position (grounded-lateral RFC D2).
-        research: Deep tier opt-in (RFC D3/D4): each persona is asked to ground
-              claims with web evidence *when its runtime exposes web tools*, and
-              to emit the machine-readable evidence block the citation gate
-              parses. Purely additive prompt contract — a runtime without web
-              tools ignores it and degrades to opinion-only, never errors.
+        mode: ``"unstuck"`` (default) or ``"decision"`` — selects the persona
+              task block (see lateral_decision.py, grounded-lateral RFC D2).
+        research: Deep tier opt-in (RFC D3/D4): personas ground claims in web
+              evidence when their runtime has web tools, else degrade to
+              opinion-only — never an error.
 
     Returns:
         List of SubagentPayload, one per unique persona.
@@ -1881,6 +1877,7 @@ def build_lateral_multi_subagent(
     Raises:
         ValueError: If personas empty, unknown, or required fields missing.
     """
+    from ouroboros.mcp.tools import lateral_decision as _ld
     from ouroboros.resilience.lateral import LateralThinker, ThinkingPersona
 
     if not personas:
@@ -1938,51 +1935,9 @@ def build_lateral_multi_subagent(
             continue
 
         lateral = result.unwrap()
-        # Wrap the persona prompt with an explicit instruction for the
-        # subagent to produce a concrete alternative plan, not just restate.
-        if mode == "decision":
-            task_block = (
-                "## Task for you (subagent)\n"
-                f"You are thinking as the **{persona.value}** persona, advising "
-                "on a decision the user must make. The problem context above "
-                "describes the choice; the current approach is the option "
-                "currently favored (if any). Produce:\n"
-                "1. Your position: which option this persona picks, in one line.\n"
-                "2. Your strongest argument for it (2-4 bullets).\n"
-                "3. The strongest argument AGAINST your own position — steelman "
-                "the other side, do not strawman it.\n"
-                "4. Flip condition: the concrete condition under which you would "
-                "switch to another option.\n\n"
-                "Keep it tight. Your output will be weighed against other "
-                "personas advising in parallel. Be distinctive — lean hard into "
-                "your persona."
-            )
-        else:
-            task_block = (
-                "## Task for you (subagent)\n"
-                f"You are thinking as the **{persona.value}** persona. Apply the "
-                "instructions above to this specific problem. Produce:\n"
-                "1. A concrete alternative plan (3-5 bullet steps).\n"
-                "2. The single biggest assumption you challenge.\n"
-                "3. A one-line verdict: would this plan work? why/why not?\n\n"
-                "Keep it tight. Your output will be compared with 4 other personas "
-                "thinking in parallel. Be distinctive — lean hard into your persona."
-            )
-        research_block = ""
-        if research:
-            research_block = (
-                "\n\n## Evidence (deep tier)\n"
-                "If your runtime exposes web tools (WebSearch/WebFetch or "
-                "equivalent), ground every load-bearing claim in a source you "
-                "actually fetched, and end your reply with exactly one fenced "
-                "json block of the shape "
-                '{"external_sources": ["<url>", ...], "claims": '
-                '[{"claim": "<one sentence>", "source": "<url>"}, ...]}. '
-                "List only URLs you fetched in this task — never invent or "
-                "recall a URL from memory. If you have no web tools, skip this "
-                "section entirely and reason from the given context; do not "
-                "fabricate sources."
-            )
+        # Mode/evidence prompt blocks live in lateral_decision.py (#1797).
+        task_block = _ld.build_lateral_task_block(persona.value, mode)
+        research_block = _ld.build_lateral_research_block(research)
         prompt = f"{lateral.prompt}\n\n---\n\n{task_block}{research_block}"
 
         context = {
@@ -1991,9 +1946,8 @@ def build_lateral_multi_subagent(
             "current_approach": current_approach,
             "failed_attempts": list(failed_attempts),
             "mode": mode,
+            **({"research": True} if research else {}),
         }
-        if research:
-            context["research"] = True
 
         payloads.append(
             build_subagent_payload(
